+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added highlight corrupted mobs in purple color
 - Added command /shmarkplayer (marking a player with yellow color)
 - Added highlight slayer miniboss in blue color
+- Added option to hide the death messages of other players, except for players who are close to the player, inside dungeon or doing a Kuudra fight.
 
 ### Fixes
 - Fixed small bazaar Message Hider